requirements.txt
asgiref==3.8.1
attrs==24.2.0
cattrs==23.2.3
Django==5.0.7
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-schema-graph==3.1.0
exceptiongroup==1.2.2
phonenumberslite==8.13.42
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
